kafka_streams/requirements.txt
faust-streaming==0.11.3
python-dotenv==1.2.1
